chore: Move ConversationParticipantsChangedSystemMessageCellDescription into its own file
